Move enum_mapper to own file to prevent circular dependency
